src/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
psycopg2-binary>=2.9.3
gunicorn==20.1.0
python-dotenv==1.0.0
Flask-Migrate==4.0.4
WTForms==3.0.1
Flask-WTF==1.1.1
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
Flask-Cors==3.0.10
requests>=2.28.1
beautifulsoup4>=4.11.1
selenium==4.9.0
webdriver-manager==3.8.6
lxml==4.9.3
geopy
redis
celery
Pillow==9.5.0